Extract latestReport to its own file
